and more sorting on block builder
